refactor(release): split release notes commands
